librbd: fix async request completion event getting lost in a race
As part of snapshot create, we currently check for the lock owner first,
if the lock is held by a different client, we forward the async request to
client that hold the lock.

In the below sequence assume client-1 is the snapshot requester and client-2 is
the lock owner which is actually handling the snap create call.
* On client-1 ImageWatcher<I>::notify_snap_create prepares the
  SnapCreatePayload and hands it over to notify_async_request
* notify_async_request will then inserts the async request to m_async_requests,
  sets up the timeout to 30 seconds, and finally notify_lock_owner with payload
* Client-2 on receiving the payload, accepts the request, checks if the request
  is new, inserts it to m_async_pending, and finally executes execute_snap_create
* In the meanwhile if the timeout hit on client-1 the request is removed from
  m_async_requests (as part of async_request_timed_out) and client-1 adds
  request to m_async_requests and resends the SnapCreatePayload to client-2
  again. This steps is repeated after each scheduled timeout.
* On the other end, client-2 on receiving an async request, checks if the
  request is new and proceeds further. When client-2 is done with snap create,
  it removes the request from m_async_pending and finally notifies client-1
  with AsyncCompletePayload
* As part of handling AsyncCompletePayload, client-1 checks if the relevant
  request exists in the m_async_requests and if the request ID match in
  m_async_requests, the request is set to complete and completely removed from
  m_async_requests.

Problem:
The race goes like this, client-1 hits the async request timeout and the
request is removed from m_async_requests. At the same time client-2 sends the
AsyncCompletePayload to client-1, and client-1 cannot find anything of that
sort in m_async_requests, which leads to client-1 missing to mark the request
to complete.

Because client-1 hasn't marked the request to complete since m_async_requests
is not found, client-1 keeps resending the request again and again. On the other
hand client-2 behaves as expected, as it just identifies the requests by ID as
already completed and just drops them on the floor.

Solution:
There is no point in removing the m_async_requests when the timeout kicked-in,
as we are anyway overriding the m_async_requests map key(AsyncRequestId) with
a new value(AsyncRequest) as part of notify_async_request everytime the request
is resent.

Fixing this will avoid the short window where the request is missing
on m_async_requests.

Fixes: https://tracker.ceph.com/issues/56417
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2102061
Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>