quincy: qa/workunits/windows: backport rbd-wnbd tests
The quincy branch doesn't currently include the rbd-wnbd python
tests. For this reason, the ceph-build scripts are fetching a
certain version from the main branch and then run each test
individually [1].

One issue is that the fio test is failing on quincy and we aren't
getting any information. It seemed to be a time out, but it still
fails even after the timeout was increased.

By backporting the tests to quincy, the tests will be invoked
through the "run-tests.ps1" script, hopefully receiving more
information.

Note that we're skipping the live resize test since this feature
isn't available in Quincy.

This will also allow us to simplify the ceph-build script.

[1] https://github.com/ceph/ceph-build/blob/b5adddec2cb03059d0bc0b5d4ffa7aac7a248fc0/scripts/ceph-windows/run_tests#L70-L85

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>