cephfs: add command "ceph fs swap"
Add a FS command that enables users to swap names of two file systems in
a single PAXOS transaction. Add an option to this command that swaps
FSCIDS along with FS names. This commands also updates the application
pool tags and fails when mirroring is enabled on either or both FSs.

Fixes: https://tracker.ceph.com/issues/58129
Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit 717031454243898071bdb254b2cadc7048057903)

Conflicts:
	doc/man/8/ceph.rst
	Conflict occurred because Ceph man page contians lesser amout of
	CephFS commands in it in Reef branch

	src/mds/FSMap.h
	* Methods like set_fscid() can be cherry-picked as it is but due
	  to different lines around, patch couldn't be applied as it is.
	* Methods like get_fscid(), get_mds_map() are absent in Reef
	  branch. They have been kept/added during this conflict
	  resolution.