cephfs: add command "ceph fs swap"
Add a FS command that enables users to swap names of two file systems in
a single PAXOS transaction. Add an option to this command that swaps
FSCIDS along with FS names. This commands also updates the application
pool tags and fails when mirroring is enabled on either or both FSs.

Fixes: https://tracker.ceph.com/issues/58129
Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit 717031454243898071bdb254b2cadc7048057903)

Conflicts:
	doc/man/8/ceph.rst
	- Conflict occured because CephFS commands are documented in
	  Quincy branch are lesser than CephFS commands documented in
	  main branch.

	src/mds/FSMap.h
	- Lines surrounding the point where method set_fscid() has to be
	  patched are different on Quincy branch compared to main
	  branch. Also, 4 methods (from main) are under conflict, out of
	  these are necessary for this commit to function. They are -
	  get_mds_map() and get_fscid().