crimson/osd/pg: remove incorrect map epoch check in PG::submit_transasction
It's entirely fine for the map_epoch to change while the op is processed
as long as none of the intervening epochs caused an interval change.  In
general, the correct way to check for an interval change is with
has_reset_since.

We don't need to do this check here at all because IOInterruptCondition
will already have performed it against the captured epoch when the
continuation resumed.

Introduced: bde76a0e1d7a
Fixes: https://tracker.ceph.com/issues/58486
Signed-off-by: Samuel Just <sjust@redhat.com>