qa: Added script for collecting Windows Event Logs
In order to collect more information for the Windows ceph-build
artifacts, the collect-event-logs.ps1 script will be run in order
to extract the event logs from the Windows client machine.

It will dump all the event logs as evtx and then convert them to
txt in order to be accessible on all platforms.

The dumped event log files can be found at %WINDIR%\EventLogs.

Signed-off-by: Stefan Chivu <schivu@cloudbasesolutions.com>