mgr/cephadm: don't provide asyncssh the ssh pub key
It doesn't actually need it for the connection, providing
it only allows it to check if it matches up with the
private key, but this breaks set-priv-key and set-pub-key
commands as they only allow setting one of the keys
at a time, always causing a mismatch. So, I think
we ought to just not provide it to asyncssh.

Fixes: https://tracker.ceph.com/issues/58622

Signed-off-by: Adam King <adking@redhat.com>