osdc: Catch exceptions thrown in CLS client decoders
And return them to the client by setting the error code and result in
the vector and returning an error from the operation as a whole.

Pass OSD failure to subsequent handlers so that, for example, in the
event of a cancellation, we don't try to decode data that isn't there.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>