mds: force replay sessionmap version
When expiring the MDLog Segments it will persist the sessionmap, but
in case the number of Segments doesn't reach up to the max segment
threshold then the expiry may not happen.

So if there is a MDS failover then the sessionmap won't be persistent
and cannot be replayed by the standby MDS when it is starting. We
should just force replay the sessionmap version just the way inotable
does.

Fixes: https://tracker.ceph.com/issues/58489
Signed-off-by: Xiubo Li <xiubli@redhat.com>