mgr/volumes: avoid returning -ESHUTDOWN back to cli
I think mgr/volumes is acting a bit fancy by checking if the
ceph-mgr daemon and plugins are under termination, thereby
returning -ESHUTDOWN to cli commands. This seems unnecessary.
None of the other mgr plugins do anything like this.

Handling of commands when the ceph-mgr is under termination
is possibly handled in one of the upper layers.

Fixes: http://tracker.ceph.com/issues/58651
Signed-off-by: Venky Shankar <vshankar@redhat.com>