qa/distros: pass --allowerasing --nobest when installing container-tools
One of the tests in the orch suite is running distro install
commands from multiple distros, causing it to first install
container-tools 3.0 and then later install container-tools,
which fails, causing the test to fail. This is sort of a bandaid
fix to getthe test to work. It will cause whatever the last
version of the package to be installed to end up being installed
(and will do so without error) which is what we want in the tests.

Fixes: https://tracker.ceph.com/issues/57771

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 3011d954986e38ae8f7c9cd027ef2a88dff9a3d8)