cephfs-top: Handle `METRIC_TYPE_NONE` fields for sorting
Currently, the sort feature in cephfs-top isn't applicable
to the fields which are of type `METRIC_TYPE_NONE` , named as
'ofiles', 'oicaps' and 'oinodes'. This commit will make the
cephfs-top sort feature to support the above mentioned fields too.
From this commit onwards, all the current fields displayed in
cephfs-top can be used for sorting.

Fixes: https://tracker.ceph.com/issues/58814
Signed-off-by: Neeraj Pratap Singh <neesingh@redhat.com>