test/c_write_operations: set RETURNVEC on expected cmpext failuer
We are sending cmpext request that return error and rval error.
In order to get reliably rval return for cmpext, we need to set RETURNVEC
flag so we record in pg log in case of reconnect during socket injection.
without that flag the re-request of the op won't have the rval set
and we will end with unmatch error.

Fixes: https://tracker.ceph.com/issues/21592
Signed-off-by: Nitzan Mordechai <nmordec@redhat.com>