rgw/curl: async curl client
adds class rgw::curl::Client that manages a curl multi handle and its
associated connection pool. member function async_perform() executes an
asynchronous http request as if by calling curl_easy_perform(). this
function is not thread-safe, and must be called from the same executor
that the Client was constructed with

unlike the existing RGWHTTPManager based on curl multi, rgw::curl::Client
uses the 'multi-socket' api which avoids the need to dedicate a background
thread to polling for curl completions. instead, it manages its own sockets
and timer with asio that run in an existing event loop (asio::io_context)

Signed-off-by: Casey Bodley <cbodley@redhat.com>