doc/rgw: remove "tertiary", link to procedure
Remove the term "tertiary zone" and replace it with "second secondary
zone" (because there is no such thing as a tertiary zone). Link to the
procedure for creating a secondary zone in a place where such a link is
helpful to the reader.

Signed-off-by: Zac Dover <zac.dover@proton.me>
(cherry picked from commit 8e273199d09a670578b0c1076bc18dbd113d42dd)