mgr/cephadm: try to avoid pull when getting container image info
only if use_repo_digest is not set.

The commit ac88200 introduced this possibility to skip pulling, but
doing this unconditionally broke a use case when one was able to have
a ceph image on a floating tag, and was able to upgrade to a new image
pushed to that tag. As using a floating tag is possible only when
use_repo_digest is enabled (the default), now skipping the pull
if use_repo_digest is disabled will not break it anymore.

Signed-off-by: Mykola Golub <mykola.golub@clyso.com>
(cherry picked from commit 4d049880a0effa78b69179694430d8a274829342)

Conflicts:
	src/pybind/mgr/cephadm/serve.py