mgr/cephadm: set OSD flags during upgrade
in this commit at least, this means the noout, noscrub,
and nodeep-scrub flags, although it's configurable with
the mgr/cephadm/osd_upgrade_flags conf option. The idea
is that during upgrade there may be situations that cause
OSDs to get marked down and out (see the linked tracker)
and so it's good practice to have these set during upgrade.
This also gets us in line with ceph-ansible, which set
these flags itself during upgrade.

Using the staggered upgrade test to test this since we
need to be upgrading from a mgr with upgrade code
that contains these changes.

Fixes: https://tracker.ceph.com/issues/56670

Signed-off-by: Adam King <adking@redhat.com>