ceph-volume: make disk.is_device follow symlinks
When calling ceph-volume on a symlink, the user intends to use the
device which is the target of the symlink. Change disk.is_device() to
use os.stat instead of os.lstat. This ensures we check whether the
target of the symlink is a block device.

This fixes adding multipath devices which can be accessed via
/dev/mapper/<id> but are symlinks to ../dm-<num>.

Fixes: https://tracker.ceph.com/issues/59375
Signed-off-by: Pierre Riteau <pierre@stackhpc.com>