mon/AuthMonitor: log when parsing caps fails
AuthMonitor::valid_caps() checks whether or not caps are valid by
checking if parsing caps is successful. The respective parsing methods
print a error message to stderr when parsing fails but none print a log
message. Let's add a log entry when parsing fails are caps are known to
be invalid.

Signed-off-by: Rishabh Dave <ridave@redhat.com>