rgw: determine bucket index oid correctly
Due to historical reasons a bucket index shard count of 0 is
equivalent to a count of 1. Recent changes have favored indicating
this with 1 rather than 0. Therefore logic that strictly tests for 0
no longer works.

Additionally this commit fixes other bits of code that strictly tests
for a bucket index shard count of 0 and tries to consolidate this
"knowledge" into a single place in the code.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>