cephadm: infer fsid when running "cephadm version"
This command needs the fsid as it routes through code that
tries to find daemons in the cluster on the host and uses the
fsid to check those daemons are in the correct cluster. This
change should be very safe as inferring an fsid only errors in
cases where there are multiple fsids present on the host in which
case we want to tell the user to provide the fsid anyway (which
is what the error message will say).

Fixes: https://tracker.ceph.com/issues/59428

Signed-off-by: Adam King <adking@redhat.com>