mgr/dashboard: fix the rbd mirroring configure check
In one-way mirroring, the condition we are checking now for configuring
the mirroring will fail because only one cluster needs to have the
mirror daemon present. Thus even if mirroring is successfuly happening
the page won't load. For now relaxing the rule until we find a better
api call to check for the status

Fixes: https://tracker.ceph.com/issues/59573
Signed-off-by: Nizamudeen A <nia@redhat.com>