cephadm: add executes_early decorator to version command
Add the executes_early decorator and apply it to the version command
such that the version command will function without the uaual
requirements and set up steps that a cephadm command needs.
This allows anyone with a binary, root or not, to check the version.

Signed-off-by: John Mulligan <jmulligan@redhat.com>
(cherry picked from commit 20e38136d7baafdf74a43c09effeffec3040f952)