mon/Elector.cc: allow pinging when quorum_mon_features is not yet populated
In `begin_peer_ping`, when we are at bootup time,
`quorum_mon_features` is not yet assigned any values
because an election have not yet taken place. To prevent
the monitor from rejecting pings at bootup time, we modify
the condition to see if we ourself (MON) supports
FEAUTURE_PINGING, if we do then we continue with
`begin_peer_ping` process. If we don't then
we just exit the function.

Fixes: https://tracker.ceph.com/issues/59564

Signed-off-by: Kamoltat <ksirivad@redhat.com>