python-common: add tests for service spec extra args handling
There were no existing tests for how the service spec(s) handled the
extra_container_args and extra_entrypoint_args values.  Add a short
parametrized test function to assert some basic properties of how the
arguments are currently handled. In particular, it asserts that the
values can appear at the top-level of the YAML and under spec.

This is in preparation for adding a more sophisticated argument
type in the future.

Signed-off-by: John Mulligan <jmulligan@redhat.com>